+++ project/Dependencies.scala
@@ -56,7 +56,7 @@ object Dependencies {
 
   val agentDependencies: Seq[ModuleID] = {
 
-    val spark3Version = "3.2.2"
+    val spark3Version = "3.3.2"
     val scala212 = "2.12.12"
     val scalatestVersion = "3.2.15"
     val specs2Version = "4.19.2"
@@ -68,12 +68,15 @@ object Dependencies {
         "com.fasterxml.jackson.module", "jackson-module-scala_" + scala212.substring(0, 4) // e.g. 2.11
       )*/
 
+    lazy val sparkCommons = "za.co.absa" % "spark-commons-spark3.3_2.12" % "0.5.0"
+
     lazy val sparkSql ="org.apache.spark" %% "spark-sql" %  spark3Version
     lazy val scalaTest = "org.scalatest" %% "scalatest" % scalatestVersion % Test
     lazy val specs2core = "org.specs2" %% "specs2-core" % specs2Version % Test
     lazy val typeSafeConfig = "com.typesafe" % "config" % typesafeConfigVersion
 
-    Seq(sparkCore, sparkSql, scalaTest, specs2core, typeSafeConfig)
+
+    Seq(sparkCore, sparkCommons, sparkSql, scalaTest, specs2core, typeSafeConfig)
 
   }}
